Refactor to use common Modal for sharing/meeting controls
