+++ lma-ai-stack/source/ui/src/components/call-list/CallList.jsx
@@ -7,13 +7,11 @@ import { Logger } from 'aws-amplify';
 
 import useCallsContext from '../../contexts/calls';
 import useSettingsContext from '../../contexts/settings';
-import useAppContext from '../../contexts/app';
 
 import mapCallsAttributes from '../common/map-call-attributes';
 import { paginationLabels } from '../common/labels';
 import useLocalStorage from '../common/local-storage';
 import { exportToExcel } from '../common/download-func';
-import { shareMeetings } from '../common/share-meeting';
 
 import {
   CallsPreferences,
@@ -46,7 +44,6 @@ const CallList = () => {
   } = useCallsContext();
 
   const [preferences, setPreferences] = useLocalStorage('call-list-preferences', DEFAULT_PREFERENCES);
-  const { currentSession, currentCredentials } = useAppContext();
 
   // prettier-ignore
   const {
@@ -78,8 +75,6 @@ const CallList = () => {
     setSelectedItems(collectionProps.selectedItems);
   }, [collectionProps.selectedItems]);
 
-  const [shareResult, setShareResult] = useState(null);
-
   /* eslint-disable react/jsx-props-no-spreading */
   return (
     <Table
@@ -87,6 +82,7 @@ const CallList = () => {
       header={
         <CallsCommonHeader
           resourceName="Meetings"
+          calls={calls}
           selectedItems={collectionProps.selectedItems}
           totalItems={callList}
           updateTools={() => setToolsOpen(true)}
@@ -96,19 +92,6 @@ const CallList = () => {
           setPeriodsToLoad={setPeriodsToLoad}
           downloadToExcel={() => exportToExcel(callList, 'Meeting-List')}
           // eslint-disable-next-line max-len, prettier/prettier
-          shareMeeting={async (recipients) => {
-            const result = await shareMeetings(
-              calls,
-              collectionProps,
-              recipients,
-              settings,
-              currentCredentials,
-              currentSession,
-            );
-            setShareResult(result);
-          }}
-          shareResult={shareResult}
-          setShareResult={setShareResult}
         />
       }
       columnDefinitions={COLUMN_DEFINITIONS_MAIN}
