+++ lma-ai-stack/source/ui/src/components/call-list/calls-table-config.js
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import React, { useState } from 'react';
+import React from 'react';
 import {
   Button,
   ButtonDropdown,
@@ -10,11 +10,6 @@ import {
   SpaceBetween,
   StatusIndicator,
   Popover,
-  Modal,
-  Form,
-  FormField,
-  Input,
-  Alert,
 } from '@awsui/components-react';
 
 import rehypeRaw from 'rehype-raw';
@@ -26,6 +21,7 @@ import { SentimentTrendIndicator } from '../sentiment-trend-icon/SentimentTrendI
 import { CategoryAlertPill } from './CategoryAlertPill';
 import { CategoryPills } from './CategoryPills';
 import { getTextOnlySummary } from '../common/summary';
+import { shareModal } from '../common/share-meeting';
 
 export const KEY_COLUMN_ID = 'callId';
 
@@ -269,29 +265,6 @@ export const CallsCommonHeader = ({ resourceName = 'Meetings', ...props }) => {
     localStorage.setItem(PERIODS_TO_LOAD_STORAGE_KEY, JSON.stringify(shardCount));
   };
 
-  const [shareMeeting, setShareMeeting] = useState(false);
-  const [meetingRecipients, setMeetingRecipients] = React.useState('');
-  const [submit, setSubmit] = useState(false);
-
-  const openShareSettings = () => {
-    setShareMeeting(true);
-  };
-
-  const closeShareSettings = () => {
-    setShareMeeting(false);
-    setMeetingRecipients('');
-    props.setShareResult(null);
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setSubmit(true);
-    console.log('Meeting Recipients: ', meetingRecipients);
-    await props.shareMeeting(meetingRecipients);
-    setMeetingRecipients('');
-    setSubmit(false);
-  };
-
   // eslint-disable-next-line
   const periodText =
     TIME_PERIOD_DROPDOWN_ITEMS.filter((i) => i.count === props.periodsToLoad)[0]?.text || '';
@@ -316,58 +289,7 @@ export const CallsCommonHeader = ({ resourceName = 'Meetings', ...props }) => {
             loading={props.loading}
             onClick={() => props.downloadToExcel()}
           />
-          <Button
-            iconName="share"
-            variant="normal"
-            loading={props.loading}
-            onClick={openShareSettings}
-            disabled={props.selectedItems.length === 0}
-          />
-          <Modal
-            onDismiss={closeShareSettings}
-            visible={shareMeeting}
-            footer={
-              <form
-                onSubmit={(e) => {
-                  e.preventDefault();
-                  handleSubmit(e);
-                }}
-              >
-                <Form
-                  actions={
-                    <SpaceBetween direction="horizontal" size="xs">
-                      <Button formAction="none" onClick={closeShareSettings}>
-                        Close
-                      </Button>
-                      <Button
-                        variant="primary"
-                        disabled={submit || !meetingRecipients.trim()}
-                        onclick={(e) => {
-                          e.preventDefault();
-                          handleSubmit(e);
-                        }}
-                      >
-                        Submit
-                      </Button>
-                    </SpaceBetween>
-                  }
-                >
-                  <FormField>
-                    <Input value={meetingRecipients} onChange={(event) => setMeetingRecipients(event.detail.value)} />
-                  </FormField>
-                  <Alert type="info" visible={props.shareResult}>
-                    {props.shareResult}
-                  </Alert>
-                </Form>
-              </form>
-            }
-            header={<h3>Share Meeting</h3>}
-          >
-            You are sharing&#xA0;
-            {props.selectedItems.length}
-            {props.selectedItems.length === 1 ? ' meeting' : ' meetings'}
-            &#x2e; Enter a comma separated list of email addresses.
-          </Modal>
+          {shareModal(props)}
         </SpaceBetween>
       }
       {...props}
