+++ lma-ai-stack/source/ui/src/components/call-panel/CallPanel.jsx
@@ -6,11 +6,6 @@ import {
   Badge,
   Box,
   Button,
-  Modal,
-  Form,
-  FormField,
-  Alert,
-  Input,
   ButtonDropdown,
   ColumnLayout,
   Container,
@@ -49,7 +44,7 @@ import useAppContext from '../../contexts/app';
 import awsExports from '../../aws-exports';
 import { downloadTranscriptAsExcel, downloadTranscriptAsText, exportToTextFile } from '../common/download-func';
 import useCallsContext from '../../contexts/calls';
-import { shareMeetings } from '../common/share-meeting';
+import { shareModal } from '../common/share-meeting';
 
 const logger = new Logger('CallPanel');
 
@@ -151,42 +146,11 @@ const CallSummary = ({ item }) => {
     }
   };
 
-  const { currentSession, currentCredentials } = useAppContext();
-  const { settings } = useSettingsContext();
   const { calls } = useCallsContext();
-
-  const [share, setShare] = useState(false);
-  const [meetingRecipients, setMeetingRecipients] = React.useState('');
-  const [submit, setSubmit] = useState(false);
-  const [shareResult, setShareResult] = useState(null);
-  const collectionProps = {
+  const props = {
+    calls,
     selectedItems: [item],
-  };
-
-  const openShareSettings = () => {
-    setShare(true);
-  };
-
-  const closeShareSettings = () => {
-    setShare(false);
-    setMeetingRecipients('');
-    setShareResult(null);
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setSubmit(true);
-    const result = await shareMeetings(
-      calls,
-      collectionProps,
-      meetingRecipients,
-      settings,
-      currentCredentials,
-      currentSession,
-    );
-    setMeetingRecipients('');
-    setShareResult(result);
-    setSubmit(false);
+    loading: false,
   };
 
   return (
@@ -217,52 +181,7 @@ const CallSummary = ({ item }) => {
                   <Icon name="download" variant="primary" />
                 </ButtonDropdown>
               )}
-              <Button iconName="share" variant="normal" onClick={openShareSettings} />
-              <Modal
-                onDismiss={closeShareSettings}
-                visible={share}
-                footer={
-                  <form
-                    onSubmit={(e) => {
-                      e.preventDefault();
-                      handleSubmit(e);
-                    }}
-                  >
-                    <Form
-                      actions={
-                        <SpaceBetween direction="horizontal" size="xs">
-                          <Button formAction="none" onClick={closeShareSettings}>
-                            Close
-                          </Button>
-                          <Button
-                            variant="primary"
-                            disabled={submit || !meetingRecipients.trim()}
-                            onclick={(e) => {
-                              e.preventDefault();
-                              handleSubmit(e);
-                            }}
-                          >
-                            Submit
-                          </Button>
-                        </SpaceBetween>
-                      }
-                    >
-                      <FormField>
-                        <Input
-                          value={meetingRecipients}
-                          onChange={(event) => setMeetingRecipients(event.detail.value)}
-                        />
-                      </FormField>
-                      <Alert type="info" visible={shareResult}>
-                        {shareResult}
-                      </Alert>
-                    </Form>
-                  </form>
-                }
-                header={<h3>Share Meeting</h3>}
-              >
-                You are sharing&#xA0; 1 meeting &#x2e; Enter a comma separated list of email addresses.
-              </Modal>
+              {shareModal(props)}
             </SpaceBetween>
           }
         >
